# Refactor the "order" state in the block editor reducer to use a map instead of a plain object