Similar to #46204 and #46146

Following a similar approach to the previous PRs, I've added a test to simulate what happens when you click the block mover to move a block up/down.

And while doing so, I tried to see how refactoring the order state from plain objects to maps could impact the performance. Here are the results I got:

Before:

<img width="492" alt="Screenshot 2022-12-01 at 10 31 23 AM" src="https://user-images.githubusercontent.com/272444/205017600-117ff341-6b54-4b02-b054-10a27390f044.png">

After:

<img width="485" alt="Screenshot 2022-12-01 at 10 30 37 AM" src="https://user-images.githubusercontent.com/272444/205017580-fb2b0522-22f9-4fd6-8ae5-de3a284fbcac.png">

As you can see:

 - The "update block" metric is stable (no impact on this PR)
 - The "hit enter to create a new block" metric improved from 378ms to 271ms
 - The "move block" metric metric improved from 62ms to 23ms

I think we can improve these metrics further if also refactor the "tree" (and other) state to maps on follow-up PRs.